refactor: extract generic utils from DappBrowser into framework
